# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flight Radar - Laravel",
        "description": "This is a Rest API built with Laravel for tracking personal flight hidstory",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/api/flights/{id}/notes": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Get all notes from sepcific a specific flight",
                "operationId": "66469e988eafcae3280ba0c479cde490",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Flight ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve notes"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Note"
                ],
                "summary": "Get a sepcific note",
                "operationId": "9b7eded99ae8449756cd089aa71078fc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve note"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Note"
                ],
                "summary": "Update an existing note",
                "operationId": "c7e6224fadd6d82b81ded7615279a15c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Note payload format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of Note",
                                        "type": "string"
                                    },
                                    "body": {
                                        "description": "Write the note",
                                        "type": "string"
                                    },
                                    "flight_id": {
                                        "description": "Flight id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated note"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve note"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Note"
                ],
                "summary": "Delete a specific note from user's list of notes",
                "operationId": "231822d83675b5b9446d8b0f6f4d3959",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted note"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve note"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/notes": {
            "post": {
                "tags": [
                    "Note"
                ],
                "summary": "Create a new note",
                "operationId": "e28ffd9dcbbcb4e1b04a525ab9c13a01",
                "requestBody": {
                    "description": "Note payload format",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Title of Note",
                                        "type": "string"
                                    },
                                    "body": {
                                        "description": "Write the note",
                                        "type": "string"
                                    },
                                    "flight_id": {
                                        "description": "Flight id",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created note"
                    },
                    "401": {
                        "description": "Token is invalid"
                    },
                    "404": {
                        "description": "Unable to retrieve note"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}